add retries mechanism

+++ src/main/java/dev/crashteam/styx/service/web/AdvancedConversationService.java
@@ -1,16 +1,20 @@
 package dev.crashteam.styx.service.web;
 
 import dev.crashteam.styx.exception.OriginalRequestException;
+import dev.crashteam.styx.exception.ProxyForbiddenException;
 import dev.crashteam.styx.exception.ProxyGlobalException;
 import dev.crashteam.styx.model.proxy.ProxyInstance;
+import dev.crashteam.styx.model.request.RetriesRequest;
 import dev.crashteam.styx.model.web.ErrorResult;
 import dev.crashteam.styx.model.web.ProxyRequestParams;
 import dev.crashteam.styx.model.web.Result;
 import dev.crashteam.styx.service.proxy.CachedProxyService;
 import dev.crashteam.styx.util.AdvancedProxyUtils;
 import io.netty.handler.proxy.ProxyConnectException;
+import io.netty.handler.timeout.ReadTimeoutException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.UnsupportedMediaTypeException;
 import org.springframework.web.reactive.function.client.ClientResponse;
@@ -18,7 +22,7 @@ import reactor.core.publisher.Mono;
 
 import java.time.Duration;
 import java.util.Objects;
-import java.util.Optional;
+import java.util.UUID;
 
 @Service
 @Slf4j
@@ -27,14 +31,16 @@ public class AdvancedConversationService {
 
     private final WebClientService webClientService;
     private final CachedProxyService proxyService;
+    private final RetriesRequestService retriesRequestService;
 
     public Mono<Result> getProxiedResult(ProxyRequestParams params) {
+        String requestId = UUID.randomUUID().toString();
         return proxyService.getRandomProxy(0L)
                 .hasElement()
                 .flatMap(hasElement -> {
                     if (hasElement) {
                         return proxyService.getRandomProxy(params.getTimeout())
-                                .flatMap(proxy -> getProxiedResponse(params, proxy));
+                                .flatMap(proxy -> getProxiedResponse(params, proxy, requestId));
                     } else {
                         return getWebClientResponse(params)
                                 .delaySubscription(Duration.ofMillis(params.getTimeout()));
@@ -43,17 +49,18 @@ public class AdvancedConversationService {
                 .onErrorResume(Objects::nonNull, e -> {
                     log.error("Unknown error", e);
                     return Mono.just(ErrorResult.unknownError(params.getUrl(), e));
-                });
+                }).doOnSuccess(result -> retriesRequestService.deleteIfExistByRequestId(requestId).subscribe());
 
     }
 
-    private Mono<Result> getProxiedResponse(ProxyRequestParams params, ProxyInstance proxy) {
+    private Mono<Result> getProxiedResponse(ProxyRequestParams params, ProxyInstance proxy, String requestId) {
         log.info("Sending request via proxy - [{}:{}]. URL - {}, HttpMethod - {}. Proxy source - {}, Bad proxy points - {}",
                 proxy.getHost(), proxy.getPort(),
                 params.getUrl(), params.getHttpMethod(), proxy.getProxySource().getValue(), proxy.getBadProxyPoint());
         return webClientService.getProxiedWebclientWithHttpMethod(params, proxy)
                 .retrieve()
                 .onStatus(httpStatus -> !httpStatus.is2xxSuccessful(), this::getMonoError)
+                .onStatus(httpStatus -> httpStatus.equals(HttpStatus.FORBIDDEN), this::getForbiddenError)
                 .toEntity(Object.class)
                 .map(response -> Result.success(response.getStatusCodeValue(), params.getUrl(), response.getBody(),
                         params.getHttpMethod()))
@@ -67,14 +74,29 @@ public class AdvancedConversationService {
                     if (e.getCause() instanceof UnsupportedMediaTypeException) {
                         return Mono.just(ErrorResult.unknownError(params.getUrl(), e));
                     }
-                    proxyService.deleteByHashKey(proxy);
-                    log.error("PROXY ERROR ", Optional.ofNullable(e.getCause()).orElse(e));
-                    return connectionErrorResult(e, params);
+                    return retriesRequestService.existsByRequestId(requestId)
+                            .flatMap(exist -> {
+                                if (exist) {
+                                    return retriesRequestService.findByRequestId(requestId);
+                                } else {
+                                    RetriesRequest retriesRequest = new RetriesRequest();
+                                    retriesRequest.setRequestId(requestId);
+                                    retriesRequest.setRetries(3);
+                                    return Mono.just(retriesRequest);
+                                }
+                            }).flatMap(retriesRequest -> {
+                                retriesRequest.setRetries(retriesRequest.getRetries() - 1);
+                                if (retriesRequest.getRetries() == 0) {
+                                    retriesRequestService.deleteByRequestId(requestId).subscribe();
+                                    proxyService.deleteByHashKey(proxy);
+                                    return Mono.just(ErrorResult.proxyConnectionError(params.getUrl(), e));
+                                }
+                                retriesRequestService.save(retriesRequest).subscribe();
+                                return connectionErrorResult(e, params, requestId);
+                            });
                 })
                 .onErrorResume(throwable -> throwable instanceof ProxyGlobalException,
-                        e -> Mono.just(ErrorResult.unknownError(params.getUrl(), e)))
-                .onErrorResume(throwable -> throwable instanceof ProxyConnectException,
-                        e -> Mono.just(ErrorResult.proxyConnectionError(params.getUrl(), e)));
+                        e -> Mono.just(ErrorResult.unknownError(params.getUrl(), e)));
 
     }
 
@@ -100,14 +122,14 @@ public class AdvancedConversationService {
                         });
     }
 
-    private Mono<Result> connectionErrorResult(Throwable e, ProxyRequestParams params) {
+    private Mono<Result> connectionErrorResult(Throwable e, ProxyRequestParams params, String requestId) {
         log.error("Trying to send request with another random proxy. Exception - " + e.getMessage());
         return proxyService.getRandomProxy(0L)
                 .hasElement()
                 .flatMap(hasElement -> {
                     if (hasElement) {
                         return proxyService.getRandomProxy(params.getTimeout()).flatMap(proxy ->
-                                getProxiedResponse(params, proxy));
+                                getProxiedResponse(params, proxy, requestId));
                     } else {
                         return getWebClientResponse(params)
                                 .delaySubscription(Duration.ofMillis(params.getTimeout()));
@@ -122,6 +144,17 @@ public class AdvancedConversationService {
                     return Mono.error(new ProxyGlobalException(e.getMessage(), e));
                 })
                 .flatMap(body -> Mono.error(new OriginalRequestException("Proxy request error", body,
+                        response.rawStatusCode())));
+
+    }
+
+    private Mono<? extends Throwable> getForbiddenError(ClientResponse response) {
+        return response.bodyToMono(String.class)
+                .onErrorResume(Objects::nonNull, e -> {
+                    log.error("Unknown error", e);
+                    return Mono.error(new ProxyGlobalException(e.getMessage(), e));
+                })
+                .flatMap(body -> Mono.error(new ProxyForbiddenException("Proxy request forbidden error", body,
                         response.rawStatusCode())));
 
     }
